pyproject.toml
[project]
name = "hackathon-flex"
version = "0.1.0"
description = "B2B routing bridge: Telegram + Email + Phone → Bitrix24 → Telegram replies"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "telethon>=1.36",
    "aioimaplib>=1.0",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
    "faster-whisper>=1.0.3",
    "openai>=1.30.0",
    "pyrogram>=2.0.106",
    "tgcrypto>=1.2.5",
]

[project.scripts]
phone-recording = "phone_recording.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [".", "phone_recording"]
